:recycle: move server & component initialization code to a module
