+++ src/instance.ts
@@ -7,8 +7,7 @@ import { robloxLogObject, robloxActionObject } from "./types/robloxTypes";
 import NetworkComponent from "./components/networkComponent";
 import TerminalComponent from "./components/terminalComponent";
 
-import { spawn } from "child_process";
-import * as path from "path"
+import InitializeServer from "./modules/initializeServer"
 
 export class ExtensionInstance {
 	branch: branch = branch.development;
@@ -18,7 +17,7 @@ export class ExtensionInstance {
 	context: ExtensionContext;
 
 	networkComponent?: NetworkComponent;
-	terminalComponent: TerminalComponent;
+	terminalComponent?: TerminalComponent;
 
 	onExtensionActivated: Signal<() => void> = new Signal();
 	onExtensionDeactivated: Signal<() => void> = new Signal();
@@ -29,32 +28,10 @@ export class ExtensionInstance {
 	constructor(context: ExtensionContext) {
 		this.context = context;
 
-		const controlInstance = spawn("node", [ 
-			path.join(this.context.extensionPath, "components", "controlComponent.js")
-		])
+		InitializeServer(this).then(() => {
+			this.terminalComponent = new TerminalComponent(this);
 
-		this.terminalComponent = new TerminalComponent(this);
- 
-		this.onExtensionActivated.connect(() => console.log("Plugin Active!"));
-
-		controlInstance.on('close', (exitStatus) => {
-			exitStatus 
-			? 
-				console.log("controlInstance :: Instance shut down improperly with an error! Abort.")
-			: 
-				console.log("controlInstance :: Shut down API server.")
-		})
-
-		controlInstance.stderr.on('data', (err) => {
-			console.log("controlInstance :: ERROR: ", err)
-		})
-
-		controlInstance.stdout.on('data', (stream) => {
-			console.log("controlInstance :: ", stream)
-		})
-
-		this.onExtensionDeactivated.connect(() => {
-			controlInstance.kill(0)
+			this.onExtensionActivated.connect(() => console.log("Plugin Active!"));
 		})
 	}
 }
